wifi: move MacLowTransmissionParameters class to a separate file
