+++ src/wifi/wscript
@@ -24,6 +24,7 @@ def build(bld):
         'model/wifi-mac-header.cc',
         'model/wifi-mac-trailer.cc',
         'model/mac-low.cc',
+        'model/mac-low-transmission-parameters.cc',
         'model/wifi-mac-queue.cc',
         'model/mac-tx-middle.cc',
         'model/mac-rx-middle.cc',
@@ -167,6 +168,7 @@ def build(bld):
         'model/mac-tx-middle.h', 
         'model/mac-rx-middle.h', 
         'model/mac-low.h',
+        'model/mac-low-transmission-parameters.h',
         'model/originator-block-ack-agreement.h',
         'model/ctrl-headers.h',
         'model/block-ack-agreement.h',
